requirements.txt
pinecone>=7.0.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
python-dotenv>=1.0.1
numpy>=1.26.0
torch>=2.2.0
requests>=2.32.0